Uses \n as header values separator
With commas as a separator, we cant `split(',')` the value to get the
multiple `set-cookie` headers because "," can be present in cookie
value or expires attribute.

Probably the correct thing to do is re-use [the standard HttpHeaders
class](https://api.flutter.dev/flutter/dart-io/HttpHeaders-class.html),
but it is a large refactor than I want to do now. :upside_down_face: